% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/android_codenames.R
\docType{data}
\name{android.CODENAMES}
\alias{android.CODENAMES}
\title{An ordered vector containing Android Codenames}
\format{An object of class \code{character} of length 13.}
\usage{
android.CODENAMES
}
\description{
An ordered vector containing Android Codenames
}
\keyword{android,}
\keyword{version}
