% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/android_map_version_number_to_codename.R
\name{android.map_version_number_to_codename}
\alias{android.map_version_number_to_codename}
\title{Map Android version number to Android release name}
\usage{
android.map_version_number_to_codename(version_string)
}
\arguments{
\item{version_string}{A string indicating the version number to be mapped to a release name}
}
\description{
This function maps a given version number to a release name.  This is very 
specific to only official Android releases. See \href{https://source.android.com/setup/build-numbers}{Android: Codenames, Tags, and Build Numbers}
For example: `6.0` becomes: `marshmallow`
}
\examples{
android_extract_version_number("4.4")
[1] "kitkat"
}
\keyword{android,}
\keyword{version}
