% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/android_extract_version_number.R
\name{android_extract_version_number}
\alias{android_extract_version_number}
\title{Extract Android Version Number}
\usage{
android_extract_version_number(str)
}
\arguments{
\item{str}{The string containting the version info to be extracted}
}
\description{
This function attempts to extract a version number from string data that
has other information about the version.  
For example: `Android MarshMallow 6.0 Os For MaPan` becomes: `6.0`
}
\examples{
android_extract_version_number("Android MarshMallow 6.0 Os For MaPan")
}
\keyword{android,}
\keyword{version}
